% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{bootstrap.elasticities.old}
\alias{bootstrap.elasticities.old}
\title{(OLD) Calculate 10,000 replicates of the entire set of bilateral elasticities}
\usage{
bootstrap.elasticities.old(ctry)
}
\arguments{
\item{ctry}{Country experiencing a change in TFP.}
}
\value{
A dataframe with 10,000 replicates of the 86^2 = 7,396
domestic and bilateral elasticities. There are four columns: i
(origin of TFP change), k (trading partner or domestic market
when i = k), the elasticity e.ik---all with identical
descriptions to those in
\link[TFPtoCO2]{emission_data.ave}---and 'boot', that
indexes the bootstrap replicates.
}
\description{
This function is not used because it is very slow, but I am keeping
it for future reference and, if necessary, ensuring that the
function 'bootstrap.elasticities()' is working well.
}
\details{
This function is necessary because the dataframe storing the 10,000
elasticities is too large (~700 MB) to be stored in GitHUB (max 100
MB). The function calculates the bilateral elasticities for the
country passed from the function \link[TFPtoCO2]{get.responses}. It
uses the bootstrap regression parameters from wild cluster
bootstrap using the function 'clusterSEs::cluster.wild.plm'. The
bootstrapped parameters and the competition indices needed by the
function are in the ./data folder.
}
