% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{bootstrap.elasticities}
\alias{bootstrap.elasticities}
\title{Calculate 10,000 replicates of the entire set of bilateral elasticities}
\usage{
bootstrap.elasticities()
}
\value{
A dataframe with 10,000 replicates of the 86^2 = 7,396
domestic and bilateral elasticities. There are four columns: i
(origin of TFP change), k (trading partner or domestic market
when i = k), the elasticity e.ik---all with identical
descriptions to those in
\link[TFPtoCO2]{emission_data.ave}---and 'boot', that
indexes the bootstrap replicates.
}
\description{
This function is necessary because the dataframe storing the 10,000
elasticities is too large (~700 MB) to be stored in GitHUB (max 100
MB). The function takes no argument. It uses the bootstrap
regression parameters from wild cluster bootstrap using the
function 'clusterSEs::cluster.wild.plm' (see
./programs/AJAE_replication//do.rev2.R). The bootstrapped parameters
and the competition indices needed by the function are in the
./data folder.
}
