% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get.changes.in.cropland.and.emissions}
\alias{get.changes.in.cropland.and.emissions}
\title{Calculate changes in domestic and foreign cropland and emissions
for a given percentage change in agricultural TFP.}
\usage{
get.changes.in.cropland.and.emissions(perc.changes.in.cropland, physicaldata)
}
\arguments{
\item{perc.changes.in.cropland}{Dataframe obtained from
\link[TFPANDCO2]{get.perc.changes.in.cropland}}

\item{physicaldata}{Dataframe with cropland ha and emissions.}
}
\value{
A dataframe with the change in domestic and foreign
cropland and emissions for a given percentage change in
agricultural TFP.
}
\description{
Calculate changes in domestic and foreign cropland and emissions
for a given percentage change in agricultural TFP.
}
