% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get.perc.changes.in.cropland}
\alias{get.perc.changes.in.cropland}
\title{Calculate percentage changes in domestic and foreign cropland for a
given percentage change in agricultural TFP.}
\usage{
get.perc.changes.in.cropland(ctry, tfp.shock, elasticities)
}
\arguments{
\item{ctry}{Country experiencing a change in TFP.}

\item{tfp.shock}{\% change in agricultural TFP.}

\item{elasticities}{Dataframe with bilateral elasticities.}
}
\value{
A dataframe with the percentage in domestic and foreign
cropland for a given percentage change in agricultural TFP.
}
\description{
This function can be used to get percentage changes in domestic and
foreign cropland for bootstrapped elasticities.
}
