% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get.resp}
\alias{get.resp}
\title{Wrap the two functions above for a more concise use in scripts:}
\usage{
get.resp(ctry, tfp.shock, elasticities, physicaldata)
}
\description{
Wrap the two functions above for a more concise use in scripts:
}
