% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{bootstrap.elasticities}
\alias{bootstrap.elasticities}
\title{Calculate 10,000 replicates of the entire set of bilateral elasticities}
\usage{
bootstrap.elasticities(ctry)
}
\arguments{
\item{ctry}{Country experiencing a change in TFP.}
}
\value{
A dataframe with 10,000 replicates of the 86^2 = 7,396
domestic and bilateral elasticities. There are four columns: i
(origin of TFP change), k (trading partner or domestic market
when i = k), the elasticity e.ik---all with identical
descriptions to those in
\link[TFPtoCO2]{emission_data.ave}---and 'boot', that
indexes the bootstrap replicates.
}
\description{
The function calculates the 10,000 replicates of each bilateral
elasticities of cropland change with respect to changes in TFP in a
given country. The function is internally called by the function
\link[TFPtoCO2]{get.responses}. The function relies in a stored
dataset with 10,000 bootstrap replicates (wild cluster bootstrap
using the function 'clusterSEs::cluster.wild.plm') of the
regression parameters in Villoria (2019).  The bootstrapped
parameters and the competition indices needed by the function are
in the ./data folder and are distributed with the function.
}
\references{
Villoria, Nelson B. 2019. "Technology Spillovers and
Land Use Change: Empirical Evidence from Global Agriculture."
American Journal of Agricultural Economics 101 (3):
870–93. https://doi.org/10.1093/ajae/aay088.
(\href{https://onlinelibrary.wiley.com/doi/full/10.1093/ajae/aay088
}{link})
}
