% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get.responses}
\alias{get.responses}
\title{Calculate changes in domestic and foreign cropland and emissions
for a given percentage change in domestic agricultural TFP.}
\usage{
get.responses(ctry, tfp.shock = 1, boot = NULL)
}
\arguments{
\item{ctry}{Country experiencing a change in TFP.}

\item{tfp.shock}{\% change in agricultural TFP.}

\item{boot}{Set boot = TRUE for using bootstrapped elasticities,
otherwise the point estimates are used.}
}
\value{
A dataframe with the domestic and foreign change in crop
area (in \% and in ha) as well as emissions (in gigatons of CO2
equivalent) from changes in cropland in countries indexed by k
driven by a given percentage change in agricultural TFP in
country i. If boot = "yes", a field 'boot' is included for
identifying the bootstrap replicates.
}
\description{
This function can be used with either the point-estimates or the
bootstrapped elasticities. The point estimates are documented in
\link[TFPtoCO2]{elasticities}. The bootstrapped elasticities
(10,000 replicates) are calculated by the function
\link[TFPtoCO2]{bootstrap.elasticities}. The elasticities are
combined with data on agricultural land and emission factors,
documented in \link[TFPtoCO2]{emission_data.ave}.
}
